AIMD_Phase2/requirements.txt
# Phase 2 RAG — Pinned Dependencies
# Install: pip install -r requirements.txt

# PDF parsing
PyMuPDF==1.24.5

# Embeddings
sentence-transformers==3.0.1
torch==2.3.1

# Vector search
faiss-cpu==1.8.0

# LLM API
anthropic==0.34.0

# Data / utilities
numpy==1.26.4
pandas==2.2.2

# Evaluation (optional but recommended)
# If you want to run RAGAS for additional metrics:
# ragas==0.1.14
# langchain==0.2.16
